requirements.txt
torch>=2.2
torchvision>=0.17
torchaudio>=2.2
opencv-python
pillow
numpy
pandas
scikit-learn
matplotlib
tqdm
albumentations
grad-cam
pydantic
pyyaml
gradio>=4.0
onnx
onnxruntime
seaborn
rich
ruff
black
isort
pre-commit
pytest
